PP for MoE (#160)

+++ dolomite_engine/hf_models/modeling_utils/mlp_blocks/moe/base.py
@@ -312,9 +312,11 @@ class MoE(nn.Module):
         acc_probs = probs.sum(0)
 
         if topk_idxs.is_cuda and is_cute_kernels_available() and self.is_hopper_or_newer_gpu:
-            freq = continuous_count_cute(x=topk_idxs.flatten(), size=num_experts).to(dtype=logits.dtype)
+            freq = continuous_count_cute(x=topk_idxs.flatten(), size=num_experts)
         else:
-            freq = bincount(topk_idxs.flatten(), minlength=num_experts).to(dtype=logits.dtype)
+            freq = bincount(topk_idxs.flatten(), minlength=num_experts)
+
+        freq = freq.float()
 
         if ProcessGroupManager.is_initialized() and ProcessGroupManager.get_data_parallel_world_size() > 1:
             freq = all_reduce(freq, reduceOp="sum", group=ProcessGroupManager.get_data_parallel_group())
@@ -324,4 +326,4 @@ class MoE(nn.Module):
 
         loss = switch_loss + 0.1 * z_loss
 
-        return loss
+        return loss.type_as(logits)
